+++ ChangeLog.rst
@@ -4,7 +4,8 @@
 Note worthy changes
 -------------------
 
-- ...
+- Shopify: if ``email_verified`` is present in the user payload, it will be used
+  to mark the email address retrieved as verified accordingly.
 
 
 65.13.1 (2025-11-20)